LayoutTests/platform/blackberry/editing/pasteboard/paste-list-002-expected.txt

CONSOLE MESSAGE: line 0: Wrong selection position on after paste.
Copy/pasting list items in a list. This test should be run with DRT for pasteboard access.

FAIL

alpha
beta
gamma
delta
